refactor(#2548 W4 PR-b): quickstart.rs simplification — 2040 -> 1100 LOC
quickstart.rs is the interactive setup wizard, called exactly once from
main.rs (the sole caller, confirmed by a full-repo sweep for
`quickstart::` references — unchanged). Wizard user-visible flow is
byte-for-byte unchanged; all 51 existing tests pass unmodified, zero
new behavior.

Three changes, all structural/mechanical, none behavioral:

1. Split the inline `#[cfg(test)] mod tests { .. }` (943 lines) into a
   sibling `src/quickstart/tests.rs` — the file already used this exact
   pattern for `review_repro_panic_io_extra`, so this converges onto an
   established precedent rather than introducing a new one. This alone
   accounts for the bulk of the reduction (2040 -> ~1103 lines).
2. Extracted `run_async()` — three near-identical
   `tokio::runtime::Builder::new_current_thread().enable_all().build()?`
   + `block_on` call sites (verify_bot, verify_bot_is_admin,
   detect_group) collapse to one helper.
3. Extracted `read_existing_telegram_state()` — `run()` and
   `run_unattended()` each independently re-read `.env`'s token and
   fleet.yaml's group_id with byte-identical logic; now one function,
   one place to get it right.

Net: 2040 -> 1100 LOC (~46%). Short of the spike's "more than half"
estimate — on close reading, the remaining ~1100 lines are dense,
behavior-guardrail code, not padding: nearly every function traces to a
specific historical fix (Sprint 56 Track H1-H4 / #525 items 1,3,4,6,7,
12-17, #2207 A1/B, #2178, #2005, #523). I found real, safe duplication
(above) but not more without either merging visibly-distinct branches
(interactive vs unattended have real precedence differences, not just
copy-paste) or touching tested edge-case handling. Flagging for
reviewer judgment per the task's own discipline rather than forcing a
number.

Also swept: `quickstart.rs:814`'s raw `std::fs::write(&fleet_path, ..)`
(the SOURCE-OF-TRUTH.md-flagged direct fleet.yaml write) is untouched in
mechanism — same single call, just shifted to line 811. Not fixed here
(out of scope), not expanded either.

MCP surface: zero changes (this file has none).

Refs #2548 Wave 4 PR-b.

Agend-Agent: gapfix-dev3
Agend-Task: t-20260703003813342973-56872-28
Agend-Branch: refactor/2548-w4b-quickstart
